shop_api/requirements-test.txt
pytest
pytest-django
model-bakery
